+++ src/rockstor/storageadmin/views/disk_smart.py
@@ -27,7 +27,7 @@ from storageadmin.models import (Disk, SMARTInfo, SMARTAttribute,
                                  SMARTCapability, SMARTErrorLog,
                                  SMARTErrorLogSummary, SMARTTestLog,
                                  SMARTTestLogDetail, SMARTIdentity)
-from fs.btrfs import (scan_disks, btrfs_uuid,
+from fs.btrfs import (btrfs_uuid,
                       pool_usage, mount_root)
 from storageadmin.serializers import SMARTInfoSerializer
 from storageadmin.util import handle_exception
